First commit
A single triangle is rendered.

Today I learned that glDrawArrays() will produce a GL_INVALID_OPERATION
error if you accidentally enable a vertex attribute that doesn’t have a
valid associated buffer. Commenting out enableAttributeArray() calls
will seem to fix the issue, but this is a red herring—the real problem
is in the creation of the VAO.