Added grass blade instance rotations
I use a mat3 vertex attribute, which works like 3 vec3 attributes. This
is probably better than passing in an angle and computing sines and
cosines.

Another good approach would be to pass sin(theta) and cos(theta) as a
vec2, and compose them into a matrix inside the shader. I don’t know
the relative benefits of this.